pyproject.toml
[tool.poetry]
name = "project_name"
version = "0.1.0"
description = ""
authors = ["ShinnosukeUesaka <shinnosuke.uesaka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
openai = "^1.12.0"
pydantic = "^2.6.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
